% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_aA.R
\name{estimate_aA}
\alias{estimate_aA}
\title{Estimates aA and TA parameters}
\usage{
estimate_aA(temperatures, half_life, aA, TA)
}
\arguments{
\item{temperatures}{numeric vector with the tested temperatures}

\item{half_life}{numeric vector with the half-life parameters for each tested temperature, estimated using the fit_half_life() function}

\item{aA}{Starting value for the aA parameter (numeric)}

\item{TA}{Starting value for the TA parameter (numeric)}
}
\value{
A list with the two fitted parameters, i.e. aA and TA
}
\description{
Estimates aA and TA parameters
}
\keyword{internal}
