% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_thermal_dose.R
\name{model_thermal_dose}
\alias{model_thermal_dose}
\title{Fits a model to an activity vs thermal dose and fits a shape parameter (n)}
\usage{
model_thermal_dose(activity, thermal_dose, n)
}
\arguments{
\item{activity}{numeric vector with the activity associated to each thermal dose (in percentage)}

\item{thermal_dose}{numeric vector with the thermal dose for each activity.}

\item{aA}{a starting value for the shape parameter (n) to be fitted (numeric)}
}
\value{
returns a list with three elements: the value of the fitted value for the shape parameter n, a numeric vector with thermal dose values and a numeric vector with predicted activity values.
}
\description{
Fits a model to an activity vs thermal dose and fits a shape parameter (n)
}
\keyword{internal}
