% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_dose.R
\name{thermal_dose}
\alias{thermal_dose}
\title{Calculates the thermal dose}
\usage{
thermal_dose(stress_time, activity, aA, TA)
}
\arguments{
\item{activity}{numeric vector with the activity associated to each stress (in percentage)}

\item{aA}{aA parameter value estimated using the estimate_aA() function}

\item{TA}{TA parameter value estimated using the estimate_aA() function}

\item{stress}{numeric vector with the time each thermal stress was applied.}
}
\value{
numeric vector with the thermal dose associated to each temperature/time tested
}
\description{
Estimates aA and TA parameters
}
\keyword{internal}
